feat: inc and dec button will change duration +/- 1 minute.
Create duration state so that we can inc/dec 1 minute using buttons. It also should be updated when 4 of the states changes. We can update the duration state manually on 4 event handlelrs , but in here we use useEffect for this.

The only downside of the useEffect solution is that it will be rendered twice. First when 1 of the 4 states change, second if will then triggered useEffect to set the duration.